stop publishing packages (#233)
Removes the workflow for publishing nightly packages from this repo.

In 24.12, development for this project is shifting to https://github.com/rapidsai/cugraph-gnn. That repo is now fully up to date with this library's source code, as of https://github.com/rapidsai/cugraph-gnn/pull/65

## Notes for Reviewers

### Why not just archive the repo?

We're not quite ready for that yet. Other changes need to be made to RAPIDS devcontainers, integration tests, etc. for 24.12.

So this PR leaves the ability to build the project from source in place, to reduce the amount of coordination needed for those other changes.

Authors:
  - James Lamb (https://github.com/jameslamb)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/wholegraph/pull/233